+++ infra/shared/storage.ts
@@ -4,8 +4,6 @@ import * as gcp from "@pulumi/gcp";
 const config = new pulumi.Config();
 const pulumiStateBucket = config.requireSecret("pulumiStateBucket");
 const gcsBucket = config.requireSecret("gcsBucket");
-const computeSa = config.requireSecret("computeSa");
-const ownerEmail = config.requireSecret("ownerEmail");
 
 // Pulumiステート保存用バケット
 export const stateBucket = new gcp.storage.Bucket("pulumi-state", {
@@ -34,23 +32,3 @@ export const dataBucket = new gcp.storage.Bucket("app-data", {
     },
   ],
 });
-
-// オーナーアカウントにデータバケットの管理権限を付与
-export const dataBucketOwnerIam = new gcp.storage.BucketIAMMember(
-  "app-data-owner",
-  {
-    bucket: dataBucket.name,
-    role: "roles/storage.admin",
-    member: pulumi.interpolate`user:${ownerEmail}`,
-  }
-);
-
-// Cloud Runデフォルトcompute SAにデータバケットへの最低限の権限を付与
-export const dataBucketIam = new gcp.storage.BucketIAMMember(
-  "app-data-compute-sa",
-  {
-    bucket: dataBucket.name,
-    role: "roles/storage.objectUser",
-    member: pulumi.interpolate`serviceAccount:${computeSa}`,
-  }
-);
